Project_2/requirements.txt
"""
Requirements file for GRIFFIN project.
"""

# Core ML libraries
torch>=2.0.1
torchvision>=0.15.2
numpy>=1.24.3
scikit-learn>=1.3.0
pandas>=2.0.3

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.65.0

# Optional: For advanced features
# jupyter>=1.0.0              # For notebook development
# tensorboard>=2.13.0         # For advanced logging
# wandb>=0.15.0               # For experiment tracking
# mlflow>=2.5.0               # For model management

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0